// src/pages/createJob/oldChangeCorporateAccount/edit/constants/types.ts
export type TFullName = {
  title: string;
  firstName: string;
  lastName: string;
};

export type TFullNameForCorporateInfo = [
  {
    title: string;
    firstName: string;
    lastName: string;
  }
];

export type TSubAddress = {
  addressNo: string;
  building?: string;
  floor?: string;
  mooNo?: string;
  soi?: string;
  road?: string;
  tambon: string;
  amphoe: string;
  province: string;
  postalCode: string;
  country: string;
  //type?: number;
};

export type TFinancialInfo = {
  registeredCapital?: number;
  revenuePerYear?: number;
  netProFitLoss?: number;
  shareholderEquity?: number;
};

export type TRegisterBusinessAddress = {
  address: TSubAddress[];
  emailAddress?: string;
  telephone?: string;
};
export type TPlaceIncorporateAddress = {
  address: TSubAddress[];
  emailAddress?: string;
  telephone?: string;
};

export type TCorporateInfo = {
  registerId?: string;
  name: string;
  registrationNo: string;
  taxId: string;
  dateOfIncorporation: string;
  isRegisteredThailand: boolean;
  isPrimaryCountry: boolean;
  registered: string;
  isRegisteredOther?: boolean;
  primary: string;
  isPrimaryOther?: boolean;
  registeredBusiness: TRegisterBusinessAddress;
  placeofIncorporation: TPlaceIncorporateAddress;
  // financial?: TFinancialInfo;
  registeredCapital?: number;
  revenuePerYear?: number;
  netProFitLoss?: number;
  shareholderEquity?: number;
};
// ...
export type TIndividualsShareholders = {
  registerId?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  sharePercentage: number;
  types?: number;
  personalId?: string;
};

export type TBank = {
  accountType: string;
  bankName: string;
  accountNo: string;
  accountLocation: string;
  swiftCode: string;
};

export type TAttorney = {
  registerId?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  addresses: TSubAddress[];
  telephone?: string;
  email?: string;
  personalId?: string;
  types?: number;
};

export type TAuthorizePerson = {
  registerId?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  addresses: TSubAddress[];
  types?: number;
  personalId?: string;
};
export type TContactPerson = {
  registerId?: string;
  fullNames: TFullName[];
  position: string;
  division: string;
  telephone: string;
  email: string;
  personalId?: string;
};

export type TDirector = {
  registerId?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  addresses: TSubAddress[];
  types?: number;
  personalId?: string;
};

export type TJuristicsShareholders = {
  registerId?: string;
  juristicName: string;
  registrationNo: string;
  registeredCountry: string;
  sharePercentage: number;
  juristicId?: string;
};